feat: add software project and provider for reading project from org
My intention is to build project specific sub-sections in the website for my software projects. They
should have the same look, but I should have the flexibility to build very different sub-websites
for them.

For starters, we have spookfox. I want to document it on my website, but want documentation to come
from actual code. An org-file achieves that for now. I'll polish it up further as I move more of my
projects here. This is a starting point.

This commit implements only providing a standard data-structure for a project. Publishing it to what
old cl-ownpress already does is still TBD.