feat: software-projects listing page
Experimented with publishing items from within item-listing page. It don't look too bad. Maybe this
idea will work. Publishers are handed down, responsible for "eventually" creating artifacts (but not
necessarily right now), but they deterministically answer "how to access the artifact(s) I am going
to produce". With this, we wont' have to worry about any broken links; because every time there is
an internal link to access, it is produced by a publisher.

Gotta do the same for other listing pages as well now. It'll be interesting to do this in home-page,
which links `/archive`, which itself is a listing. Also with blog-posts, which link to tag-listing,
which in turn link to the post itself.