feat: resolve linked notes in a note
Finally, we can resolve all linked notes of a provided denote. I decided to always parse note-body
as html and store it as a plump-node. We can change this in future or turn it into a facade. Right
now, 100% of my usecase involve note-body to always be HTML, so I am going with simplest approach.